Extract coercers from HivePageSource
